# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_jcom-dev_zmanim_internal_dsl.CalculationStep:
    properties:
      step:
        type: string
      value:
        type: string
    type: object
  github_com_jcom-dev_zmanim_internal_dsl.ValidationError:
    properties:
      column:
        type: integer
      line:
        type: integer
      message:
        type: string
      suggestion:
        type: string
    type: object
...
  internal_handlers.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  internal_handlers.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/internal_handlers.APIError'
      meta:
        $ref: '#/definitions/internal_handlers.ResponseMeta'
    type: object
...
  internal_handlers.DSLPreviewRequest:
    properties:
      date:
        description: ISO 8601 date (YYYY-MM-DD)
        type: string
      elevation:
        description: Optional elevation in meters
        type: number
      formula:
        type: string
      latitude:
        description: Direct coordinates
        type: number
      location_id:
        description: 'Optional: city/location ID'
        type: string
      longitude:
        type: number
      timezone:
        description: e.g., "America/New_York"
        type: string
    type: object
  internal_handlers.DSLPreviewResponse:
    properties:
      breakdown:
        items:
          $ref: '#/definitions/github_com_jcom-dev_zmanim_internal_dsl.CalculationStep'
        type: array
      result:
        description: Formatted time (HH:MM:SS)
        type: string
      timestamp:
        description: Unix timestamp
        type: integer
    type: object
  internal_handlers.DSLPreviewWeekRequest:
    properties:
      elevation:
        type: number
      formula:
        type: string
      latitude:
        description: Direct coordinates
        type: number
      location_id:
        description: 'Optional: city/location ID'
        type: string
      longitude:
        type: number
      start_date:
        description: ISO 8601 date (YYYY-MM-DD)
        type: string
      timezone:
        description: e.g., "America/New_York"
        type: string
    type: object
  internal_handlers.DSLPreviewWeekResponse:
    properties:
      days:
        items:
          $ref: '#/definitions/internal_handlers.DayPreview'
        type: array
    type: object
  internal_handlers.DSLValidateRequest:
    properties:
      formula:
        type: string
    type: object
  internal_handlers.DSLValidateResponse:
    properties:
      dependencies:
        items:
          type: string
        type: array
      errors:
        items:
          $ref: '#/definitions/github_com_jcom-dev_zmanim_internal_dsl.ValidationError'
        type: array
      valid:
        type: boolean
    type: object
  internal_handlers.DayPreview:
    properties:
      date:
        description: YYYY-MM-DD
        type: string
      events:
        description: Jewish holidays, Shabbat, etc.
        items:
          type: string
        type: array
      hebrew_date:
        description: Hebrew date string
        type: string
      is_shabbat:
        type: boolean
      is_yom_tov:
        type: boolean
      result:
        description: Calculated time HH:MM:SS
        type: string
      sunrise:
        description: HH:MM:SS
        type: string
      sunset:
        description: HH:MM:SS
        type: string
    type: object
...
  internal_handlers.ResponseMeta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zmanim.com
    name: Zmanim Lab Support
  description: API for Jewish prayer times calculation and publisher management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://zmanim.com/terms
  title: Zmanim Lab API
  version: "1.0"
paths:
...
  /dsl/preview:
    post:
      consumes:
      - application/json
      description: Calculates the result of a zmanim formula for a specific date and
        location, returning the time and calculation breakdown
      parameters:
      - description: Formula, date, and location
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.DSLPreviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Calculation result with breakdown
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.DSLPreviewResponse'
              type: object
        "400":
          description: Invalid request or formula error
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
        "404":
          description: Location not found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: Preview DSL formula
      tags:
      - DSL
  /dsl/preview-week:
    post:
      consumes:
      - application/json
      description: Calculates a zmanim formula for 7 consecutive days starting from
        the specified date, including Hebrew dates and Shabbat/holiday markers
      parameters:
      - description: Formula, start date, and location
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.DSLPreviewWeekRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Weekly calculation results
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.DSLPreviewWeekResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
        "404":
          description: Location not found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: Preview DSL formula for a week
      tags:
      - DSL
  /dsl/validate:
    post:
      consumes:
      - application/json
      description: Validates a zmanim formula written in the DSL syntax, returning
        any errors and dependencies
      parameters:
      - description: Formula to validate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.DSLValidateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Validation result
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.DSLValidateResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: Validate DSL formula
      tags:
      - DSL
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token from Clerk authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"